Move PathData and related code into separate module
